+++ src/utils/config.js
@@ -1,6 +1,7 @@
 let config = {
     defaultContainerElement: null,
     defaultIconPack: 'mdi',
+    defaultIconComponent: null,
     defaultDialogConfirmText: null,
     defaultDialogCancelText: null,
     defaultSnackbarDuration: 3500,
@@ -17,11 +18,13 @@ let config = {
     defaultUnselectableDaysOfWeek: null,
     defaultTimeFormatter: null,
     defaultTimeParser: null,
+    defaultModalCanCancel: null,
     defaultModalScroll: null,
     defaultDatepickerMobileNative: true,
     defaultTimepickerMobileNative: true,
     defaultNoticeQueue: true,
-    defaultInputHasCounter: true
+    defaultInputHasCounter: true,
+    defaultUseHtml5Validation: true
 }
 
 export default config